opengl: Work around buggy booleans in Mesa
Mesa 10.2.9 and older versions are oblivious to a behavior change
in the CMP instruction on Intel CPU SandyBridge and newer.

On SandyBridge and newer ones, CMP instruction sets all bits to one
in dst register (-1) as boolean true value. Before that, only the
LSB is set to one with other bits being undefined.

Mesa 10.2.9 and older use XOR instruction on the LSB for the logical
not operator, which produces -2 as boolean value for !true.
The value is then used by SEL instruction in mix(), which compares
the value with zero and does not clear high bits before that,
selecting wrong components.

A macro MESA_BUGGY_BOOL_CMP is added to forcibly convert -1 to 1
for Mesa 10.2.9 and older before logical not result is used for
mix(). The rest of comparison operators and conditionals are safe
from this behavior.

I could not independently reproduce this behavior in a seperate
standalone problem. It is possibly because instruction generation
varies from optimization.

This behavior was fixed in Mesa upstream
2e51dc838be177a09f60958da7d1d904f1038d9c, only appearing in 10.3+.