Separate public and internal API
Several LIBFREENECT_API macros are removed from identifiers that
are no longer public. Several headers are moved to internal
directory and no longer exported.

Build for Protonect out-of-tree with public API only. This provides
a demo on how to use the public API.

Protonect will be built by default in libfreenect2, controlled with
BUILD_EXAMPLES.